+++ docs/changelog.rst
@@ -13,6 +13,7 @@ v1.2.0 - 2016-xx-xx
 - Pie charts on trends page overlap their canvas (`#136 <https://github.com/dennissiemensma/dsmr-reader/issues/136>`_).
 - 'Slumber' consumption (`#115 <https://github.com/dennissiemensma/dsmr-reader/issues/115>`_).
 - Show lowest & highest Watt peaks (`#138 <https://github.com/dennissiemensma/dsmr-reader/issues/138>`_).
+- Reset day & hour statistics when changing energy prices (`#95 <https://github.com/dennissiemensma/dsmr-reader/issues/95>`_).
 
 
 v1.1.2 - 2016-05-01